configure.py: don't delete build.ninja when rebuild is interrupted.
In commit 9cc9facbea886ea224ffcba9a89d3dd36162d70d, I fixed issue #4706.
That issue about what happens when interrupting a rebuild of build.ninja
(which happens automatically when you run "ninja" after configure.py
changed). We don't want to leave behind a half-built build.ninja,
or leave it deleted.

The solution in that commit was for configure.py to build a temporary file
(build.ninja.tmp), and only as the very last step rename it build.ninja.

Unfortunately, since that time, we added more last steps after what
used to be that very last step :-(

If this new code running after the rename takes a noticable amount of
time, and if the user is unlucky enough to interrupt it during that
time, ninja will see a modified output file (build.ninja) and a failed
rule, and will delete the output file!

So the solution is to move the rename back to being the very last step
of configure.py. There is still a theoretical chance of interrupting
configure.py between the rename and Python's exit, but this chance is
now practically zero (and also, there is no chance for configure.py to
fail on an error at that point).

Fixes #4706 (again)

Signed-off-by: Nadav Har'El <nyh@scylladb.com>