cql3: Add null and unset checks in collection validation
Validating a collection should ensure that there
are no null or unset values inside the collection.

The validation already fails in case of such values,
but it does so in an ugly way.
Length of null and unset value is negative but is
cast to unsigned size_t. Then it tries to read
a really large value and fails with marshalling error.

The new checks are a better way to handle this.

Signed-off-by: cvybhu <jan.ciolek@scylladb.com>