compaction: Introduce emergency_compaction_task
Emergency compaction's goal is to quickly reduce the number of sstables.

To achieve the goal, it
- is kicked off after flush only if there are enough sstables that
  aren't currently being compacted;
- does not serialize with other compactions;
- chooses only the smallest sstables that aren't currently being compacted;

It's a one-off best effort. If it fails for whatever reason,
either a new one will be kicked off after next flush or
regular compaction will handle the sstables. A failure will be logged,
but won't be rescheduled or reported up the call stack.