cql3: expr: stub type inference for prepare_expression
In CQL (and SQL) types flow in different directions in expression
components. In an expression

  A[:x] = :y

The type of A is known, the type of :x is derived from the type of A,
and the type of :y is derived from the type of A[:x].

Currently prepare_expression() only supports the second mode - an
expression's type is dictated by its caller via the column_specification
parameter. But this means it can only be used to evaluate the
right-hand-side of binary expressions, since the left-hand-side uses
the first mode, where the type is derived from the column, not
imposed by the caller.

To support both modes, make the column_specification parameter optional
(it is already a pointer so just accept null) and also make the returned
expression optional, to indicate failure to infer the type if the
column_specification was not given.

This patch only arranges for the new calling convention (as a new
try_prepare_expression call), it does not actually implement anything.