configure.py: reset the inline threshold for types.o to default
On my system, when compiling with LTO and PGO, our boosted inline threshold
(2500) caused types.o to compile for more than 12 hours. (I terminated the compilation at that
point).
As a temporary fix, bring the inline threshold for types.o back to Clang's
default of 225.
This is a transient change that can be reverted at any time if the problem
isn't present on a specific toolchain.