configure.py: rework the seastar_cflags variable
The name of this variable is misleading. What it really does is pass flags to
static libraries compiled by us, not just to seastar.
We will need this capability to implement LTO+PGO+BOLT in our build. We will
also need to pass linker flags, and we will need to vary those flags depending
on the build mode.

This patch splits the seastar_cflags variable into per-mode lib_cflags and
lib_ldflags variables. It shouldn't change the resulting build.ninja for now,
but will be needed by the upcoming LTO+PGO+BOLT patches.